# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fresher Resume Builder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /resumes:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List user's resumes
      responses:
        "200":
          description: List of resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"
    post:
      operationId: createResume
      tags: [resumes]
      summary: Create a new resume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeInput"
      responses:
        "201":
          description: Resume created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
  /resumes/{resumeId}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a resume with all sections
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full resume data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResumeDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateResume
      tags: [resumes]
      summary: Update resume metadata
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResumeUpdate"
      responses:
        "200":
          description: Updated resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /resumes/{resumeId}/duplicate:
    post:
      operationId: duplicateResume
      tags: [resumes]
      summary: Duplicate a resume
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
  /resumes/{resumeId}/ats-score:
    get:
      operationId: getAtsScore
      tags: [resumes]
      summary: Get ATS compatibility score and tips
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ATS score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtsScore"
  /resumes/{resumeId}/export/pdf:
    get:
      operationId: exportResumePdf
      tags: [export]
      summary: Export resume as PDF (returns base64 encoded PDF)
      parameters:
        - name: resumeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PDF export result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportResult"

  # --- Personal Info ---
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Resume:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        resumeName:
          type: string
        templateId:
          type: integer
        isPublic:
          type: boolean
        publicUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - resumeName
        - templateId
        - isPublic
        - createdAt
        - updatedAt
    ResumeInput:
      type: object
      properties:
        resumeName:
          type: string
        templateId:
          type: integer
      required:
        - resumeName
        - templateId
    ResumeUpdate:
      type: object
      properties:
        resumeName:
          type: string
        templateId:
          type: integer
        isPublic:
          type: boolean
    ResumeDetail:
      type: object
      properties:
        resume:
          $ref: "#/components/schemas/Resume"
        personalInfo:
          $ref: "#/components/schemas/PersonalInfo"
        objective:
          $ref: "#/components/schemas/Objective"
        education:
          type: array
          items:
            $ref: "#/components/schemas/Education"
        skills:
          type: array
          items:
            $ref: "#/components/schemas/Skill"
        projects:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        experience:
          type: array
          items:
            $ref: "#/components/schemas/Experience"
        certifications:
          type: array
          items:
            $ref: "#/components/schemas/Certification"
        languages:
          type: array
          items:
            $ref: "#/components/schemas/Language"
      required:
        - resume
        - education
        - skills
        - projects
        - experience
        - certifications
        - languages
    PersonalInfo:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        linkedin:
          type: ["string", "null"]
        portfolio:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
      required:
        - id
        - resumeId
        - fullName
        - email
        - phone
...
    Objective:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        summaryText:
          type: string
      required:
        - id
        - resumeId
        - summaryText
...
    Education:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        institution:
          type: string
        degree:
          type: string
        fieldOfStudy:
          type: string
        graduationYear:
          type: integer
        cgpa:
          type: string
      required:
        - id
        - resumeId
        - institution
        - degree
        - fieldOfStudy
        - graduationYear
        - cgpa
...
    Skill:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        skillName:
          type: string
        proficiencyLevel:
          type: string
          enum: [Beginner, Intermediate, Advanced, Expert]
      required:
        - id
        - resumeId
        - skillName
        - proficiencyLevel
...
    Project:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        projectTitle:
          type: string
        description:
          type: string
        technologies:
          type: string
        projectLink:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
      required:
        - id
        - resumeId
        - projectTitle
        - description
        - technologies
...
    Experience:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        company:
          type: string
        position:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        isCurrent:
          type: boolean
        responsibilities:
          type: string
      required:
        - id
        - resumeId
        - company
        - position
        - startDate
        - isCurrent
        - responsibilities
...
    Certification:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        certName:
          type: string
        issuingOrg:
          type: ["string", "null"]
        dateIssued:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
      required:
        - id
        - resumeId
        - certName
...
    Language:
      type: object
      properties:
        id:
          type: integer
        resumeId:
          type: integer
        languageName:
          type: string
        proficiency:
          type: string
          enum: [Native, Fluent, Intermediate, Basic]
      required:
        - id
        - resumeId
        - languageName
        - proficiency
...
    AtsScore:
      type: object
      properties:
        score:
          type: integer
        tips:
          type: array
          items:
            type: string
        keywordsFound:
          type: array
          items:
            type: string
        missingKeywords:
          type: array
          items:
            type: string
      required:
        - score
        - tips
        - keywordsFound
        - missingKeywords
    ExportResult:
      type: object
      properties:
        pdfBase64:
          type: string
        filename:
          type: string
      required:
        - pdfBase64
        - filename
...