# lib/api-zod/src/generated/types/education.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fresher Resume Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface Education {
  id: number;
  resumeId: number;
  institution: string;
  degree: string;
  fieldOfStudy: string;
  graduationYear: number;
  cgpa: string;
}