# lib/api-zod/src/generated/types/objective.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fresher Resume Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface Objective {
  id: number;
  resumeId: number;
  summaryText: string;
}