# lib/api-zod/src/generated/types/personalInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fresher Resume Builder API
 * OpenAPI spec version: 0.1.0
 */

export interface PersonalInfo {
  id: number;
  resumeId: number;
  fullName: string;
  email: string;
  phone: string;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  portfolio?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  photoUrl?: string | null;
}