# lib/api-zod/src/generated/types/resumeDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fresher Resume Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { Certification } from './certification';
import type { Education } from './education';
import type { Experience } from './experience';
import type { Language } from './language';
import type { Objective } from './objective';
import type { PersonalInfo } from './personalInfo';
import type { Project } from './project';
import type { Resume } from './resume';
import type { Skill } from './skill';

export interface ResumeDetail {
  resume: Resume;
  personalInfo?: PersonalInfo;
  objective?: Objective;
  education: Education[];
  skills: Skill[];
  projects: Project[];
  experience: Experience[];
  certifications: Certification[];
  languages: Language[];
}