# lib/api-zod/src/generated/types/skill.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fresher Resume Builder API
 * OpenAPI spec version: 0.1.0
 */
import type { SkillProficiencyLevel } from './skillProficiencyLevel';

export interface Skill {
  id: number;
  resumeId: number;
  skillName: string;
  proficiencyLevel: SkillProficiencyLevel;
}